ci(rendered-manifests): Rebase PR branch on each workflow run and use follow-up commits
Update workflow logic to rebase the existing PR branch and add follow-up
commits instead of deleting and recreating the branch and PR on each
run. This allows multiple commits to accumulate in the PR, making it
easier to compare changes when something looks off in the rendered
manifest and requires further adjustment in the base branch before
merging.

Specifically, this commit updates the workflow to:

- Rebase the PR branch when it already exists and is behind the base
  branch, render the manifest in a new commit, and force-push the
  changes
- Reference the commit (SHA) at which the manifest was rendered in the
  commit message, PR title and body. The PR title and body references
  are updated on each follow-up commit with the latest commit.
- Close PR and delete its branch when the generated manifest is up to
  date with the base branch, adding a comment describing so

Signed-off-by: George Thomadakis <gthomadakis@protonmail.com>